# src/swagger.yaml
openapi: 3.0.0
info:
  title: OA-Pank API
  version: 1.0.1
  description: API documentation for OA-Pank banking system with RFC 7231 compliant responses
servers:
  - url: /
    description: Current server
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: string
          example: error_code
        details:
          type: string
          example: Detailed error information
...
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: username
              message:
                type: string
                example: Username is required
              value:
                type: string
                example: ""
...
    Transaction:
      type: object
      properties:
        _id:
          type: string
          example: "60d21b4667d0d8992e610c85"
          description: Unique identifier for the transaction
        transactionId:
          type: string
          example: OAP-1620000000000-12345
          description: Human-readable transaction identifier
        accountFrom:
          type: string
          example: 31359474302
          description: Source account number
        accountTo:
          type: string
          example: 31300000000
          description: Destination account number
        amount:
          type: number
          format: float
          example: 100.00
          description: Transaction amount
        currency:
          type: string
          enum: [EUR, USD, GBP, CHF, JPY, AUD, CAD, SEK, NOK, DKK, PLN, CZK]
          example: EUR
          description: Transaction currency
        status:
          type: string
          enum: [pending, inProgress, completed, failed]
          example: completed
          description: |
            Current status of the transaction:
            - pending: Transaction is waiting to be processed
            - inProgress: Transaction is being processed
            - completed: Transaction has been successfully completed
            - failed: Transaction has failed
        type:
          type: string
          enum: [internal, external]
          example: internal
          description: |
            Type of transaction:
            - internal: Transaction between accounts in our bank
            - external: Transaction to an account in another bank
        description:
          type: string
          example: Payment for services
          description: Description or purpose of the transaction
        reference:
          type: string
          example: 4f7d6e5c-3b2a-1098-7f6e-5d4c3b2a1098
          description: Unique reference for the transaction (used for tracking)
        errorMessage:
          type: string
          nullable: true
          description: Error message if the transaction failed
        initiatedBy:
          type: string
          example: 60d21b4667d0d8992e610c85
          description: ID of the user who initiated the transaction
        signature:
          type: string
          nullable: true
          description: Digital signature for external transactions
        senderBank:
          type: string
          nullable: true
          description: Name or identifier of the sending bank
        receiverBank:
          type: string
          nullable: true
          description: Name or identifier of the receiving bank
        message:
          type: string
          example: "Transaction to Bank Name (31300000000) has been queued for processing"
          description: Additional information about the transaction status
        externalBankName:
          type: string
          example: "Reference Bank"
          description: Name of the external bank for external transactions
        createdAt:
          type: string
          format: date-time
          description: When the transaction was created
        updatedAt:
          type: string
          format: date-time
          description: When the transaction was last updated
    TransactionInput:
      type: object
      required:
        - accountFrom
        - accountTo
        - amount
        - currency
      properties:
        accountFrom:
          type: string
          example: 31359474302
          description: The source account number (must exist in our bank and belong to the authenticated user)
        accountTo:
          type: string
          example: 31300000000
          description: |
            The destination account number. Can be:
            1. An account in our bank (starts with our bank prefix and exists in our database)
            2. An account in another bank (starts with another bank's prefix)
            3. A test account (starts with our bank prefix but doesn't exist in our database)
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 100.00
          description: The amount to transfer (must be greater than 0)
        currency:
          type: string
          enum: [EUR, USD, GBP, CHF, JPY, AUD, CAD, SEK, NOK, DKK, PLN, CZK]
          example: EUR
          description: The currency of the transaction
        description:
          type: string
          example: Payment for services
          description: Optional description for the transaction
...
paths:
...
  /transactions:
    get:
      tags:
        - Transactions
      summary: Get all transactions
      description: Returns all transactions for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: query
          required: false
          schema:
            type: string
          description: Filter by account ID
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, inProgress, completed, failed]
          description: Filter by transaction status
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [internal, external]
          description: Filter by transaction type
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter by start date (YYYY-MM-DD)
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter by end date (YYYY-MM-DD)
      responses:
        "200":
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Transactions
      summary: Create transaction
      description: |
        Creates a new transaction. The system handles three types of transactions:

        1. Internal transactions - When the destination account exists in our bank
        2. External transactions - When the destination account belongs to another bank
        3. Test transactions - When the destination account has our bank prefix but doesn't exist in our database

        For test transactions, the system will create a pending transaction without verifying the account's existence.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Transaction created successfully (for internal transactions)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  transaction:
                    $ref: "#/components/schemas/Transaction"
        "202":
          description: Transaction accepted and queued for processing (for external or test transactions)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  transaction:
                    allOf:
                      - $ref: "#/components/schemas/Transaction"
                      - type: object
                        properties:
                          message:
                            type: string
                            example: "Transaction to Bank Name (31300000000) has been queued for processing"
                          externalBankName:
                            type: string
                            example: "Reference Bank"
        "400":
          description: Bad request - Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - No permission to access the account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found - Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Unprocessable Entity - Insufficient funds, unknown bank, or other business logic error
          content:
            application/problem+json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      type:
                        type: string
                        format: uri
                        example: "https://example.com/insufficient-funds"
                      title:
                        type: string
                        example: "Insufficient Funds"
                      status:
                        type: integer
                        example: 422
                      detail:
                        type: string
                        example: "Your current balance is 100.50 EUR, but the transaction requires 200.00 EUR"
                      instance:
                        type: string
                        format: uri
                        example: "/transactions"
                      available:
                        type: number
                        example: 100.50
                      availableCurrency:
                        type: string
                        example: "EUR"
                      required:
                        type: number
                        example: 200.00
                      requiredCurrency:
                        type: string
                        example: "EUR"
                      accountNumber:
                        type: string
                        example: "OAP12345678"
                      success:
                        type: boolean
                        example: false
                  - type: object
                    properties:
                      type:
                        type: string
                        format: uri
                        example: "https://example.com/unknown-bank"
                      title:
                        type: string
                        example: "Unknown Bank"
                      status:
                        type: integer
                        example: 422
                      detail:
                        type: string
                        example: "The bank with prefix 999 is not recognized"
                      instance:
                        type: string
                        format: uri
                        example: "/transactions"
                      bankPrefix:
                        type: string
                        example: "999"
                      accountNumber:
                        type: string
                        example: "99900000000"
                      success:
                        type: boolean
                        example: false
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"